added label layout recognition functionality
