+++ plugins/de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/editors/elements/GraphicalFeature.java
@@ -37,6 +37,7 @@ import de.ovgu.featureide.fm.ui.editors.FeatureConnection;
 import de.ovgu.featureide.fm.ui.editors.FeatureUIHelper;
 import de.ovgu.featureide.fm.ui.editors.IGraphicalFeature;
 import de.ovgu.featureide.fm.ui.editors.IGraphicalFeatureModel;
+import de.ovgu.featureide.fm.ui.editors.featuremodel.figures.CollapsedDecoration;
 
 /**
  * Graphical representation of an {@link IFeature} instance.
@@ -61,6 +62,8 @@ public class GraphicalFeature implements IGraphicalFeature {
 	protected Dimension dimension = new Dimension(10, 10);
 
 	private IEventListener uiObject;
+
+	private CollapsedDecoration deco;
 
 	public GraphicalFeature(IFeature correspondingFeature, IGraphicalFeatureModel graphicalFeatureModel) {
 		this.graphicalFeatureModel = graphicalFeatureModel;
@@ -268,6 +271,26 @@ public class GraphicalFeature implements IGraphicalFeature {
 			}
 		}
 		return Collections.unmodifiableList(features);
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * @see de.ovgu.featureide.fm.ui.editors.IGraphicalFeature#setCollapsedDecoration(de.ovgu.featureide.fm.ui.editors.featuremodel.figures.CollapsedDecoration)
+	 */
+	@Override
+	public void setCollapsedDecoration(CollapsedDecoration decoration) {
+		deco = decoration;
+
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * @see de.ovgu.featureide.fm.ui.editors.IGraphicalFeature#getCollapsedDecoration()
+	 */
+	@Override
+	public CollapsedDecoration getCollapsedDecoration() {
+
+		return deco;
 	}
 
 }
